Fix type conversion error in batch initialization
- Convert UInt32 to Int32 for maxTokens parameter

+++ Sources/Kuzco/LlamaInstance.swift
@@ -178,7 +178,7 @@ public class LlamaInstance {
 
             // Create batch with error handling
             do {
-                self.batchBox = LlamaBatchBox(maxTokens: self.settings.processingBatchSize)
+                self.batchBox = LlamaBatchBox(maxTokens: Int32(self.settings.processingBatchSize))
                 publishProgress(LoadUpdate(stage: .creatingContext, detail: "Token batch created."))
             } catch let error {
                 throw KuzcoError.batchCreationFailed
